@@ -7,22 +7,21 @@ name = "pydgraph"
 authors = [
   { name = "Istari Digital, Inc.", email = "dgraph-admin@istaridigital.com" },
 ]
-license = { file = "LICENSE" }
+license = "Apache-2.0"
 description = "Official Dgraph client implementation for Python"
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.9"
 classifiers = [
   'Intended Audience :: Developers',
-  'License :: OSI Approved :: Apache Software License',
   'Operating System :: OS Independent',
   'Topic :: Database',
   'Topic :: Software Development',
-  'Programming Language :: Python :: 3.7',
-  'Programming Language :: Python :: 3.8',
   'Programming Language :: Python :: 3.9',
   'Programming Language :: Python :: 3.10',
   'Programming Language :: Python :: 3.11',
   'Programming Language :: Python :: 3.12',
+  'Programming Language :: Python :: 3.13',
+  'Programming Language :: Python :: 3.14',
 ]
 dependencies = ["grpcio>=1.65.0,<2.0.0", "protobuf>=4.23.0,<7.0.0"]
 dynamic = ["version"]
