@@ -1,21 +1,24 @@
-# SPDX-FileCopyrightText: © Hypermode Inc. <hello@hypermode.com>
+# SPDX-FileCopyrightText: © 2017-2026 Istari Digital, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
 """Async stub for RPC requests using grpc.aio."""
+
+from __future__ import annotations
+
+import contextlib
+import warnings
+from typing import Any
+from urllib.parse import urlparse
 
 import grpc
 import grpc.aio
 
 from pydgraph.meta import VERSION
+from pydgraph.proto import api_pb2 as api
 from pydgraph.proto import api_pb2_grpc as api_grpc
 
-try:
-    from urllib.parse import urlparse
-except ImportError:
-    from urlparse import urlparse
-
-__author__ = "Hypermode Inc."
-__maintainer__ = "Hypermode Inc. <hello@hypermode.com>"
+__author__ = "Istari Digital, Inc."
+__maintainer__ = "Istari Digital, Inc. <dgraph-admin@istaridigital.com>"
 __version__ = VERSION
 __status__ = "development"
 
@@ -23,7 +26,12 @@ __status__ = "development"
 class AsyncDgraphClientStub:
     """Async stub for the Dgraph grpc client using grpc.aio."""
 
-    def __init__(self, addr="localhost:9080", credentials=None, options=None):
+    def __init__(
+        self,
+        addr: str = "localhost:9080",
+        credentials: grpc.ChannelCredentials | None = None,
+        options: list[tuple[str, Any]] | None = None,
+    ) -> None:
         """Initialize async client stub.
 
         Args:
@@ -38,7 +46,13 @@ class AsyncDgraphClientStub:
 
         self.stub = api_grpc.DgraphStub(self.channel)
 
-    async def login(self, login_req, timeout=None, metadata=None, credentials=None):
+    async def login(
+        self,
+        login_req: api.LoginRequest,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Response:
         """Async login operation.
 
         Args:
@@ -54,7 +68,13 @@ class AsyncDgraphClientStub:
             login_req, timeout=timeout, metadata=metadata, credentials=credentials
         )
 
-    async def alter(self, operation, timeout=None, metadata=None, credentials=None):
+    async def alter(
+        self,
+        operation: api.Operation,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Payload:
         """Async alter operation for schema changes.
 
         Args:
@@ -70,7 +90,13 @@ class AsyncDgraphClientStub:
             operation, timeout=timeout, metadata=metadata, credentials=credentials
         )
 
-    async def query(self, req, timeout=None, metadata=None, credentials=None):
+    async def query(
+        self,
+        req: api.Request,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Response:
         """Async query or mutate operation.
 
         Args:
@@ -86,7 +112,13 @@ class AsyncDgraphClientStub:
             req, timeout=timeout, metadata=metadata, credentials=credentials
         )
 
-    async def commit_or_abort(self, ctx, timeout=None, metadata=None, credentials=None):
+    async def commit_or_abort(
+        self,
+        ctx: api.TxnContext,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.TxnContext:
         """Async commit or abort operation.
 
         Args:
@@ -102,7 +134,13 @@ class AsyncDgraphClientStub:
             ctx, timeout=timeout, metadata=metadata, credentials=credentials
         )
 
-    async def check_version(self, check, timeout=None, metadata=None, credentials=None):
+    async def check_version(
+        self,
+        check: api.Check,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Version:
         """Async version check operation.
 
         Args:
@@ -118,94 +156,99 @@ class AsyncDgraphClientStub:
             check, timeout=timeout, metadata=metadata, credentials=credentials
         )
 
-    async def close(self):
-        """Close the async channel gracefully."""
-        try:
-            await self.channel.close()
-        except Exception:
-            pass  # nosec B110 - Intentionally ignore errors during channel cleanup
-
     @staticmethod
-    def parse_host(cloud_endpoint):
+    def parse_host(cloud_endpoint: str) -> str:
         """Converts any cloud endpoint to grpc endpoint.
 
-        Inserts .grpc. subdomain for Dgraph Cloud endpoints when appropriate.
         Handles IPv6, single-label hosts, and various URL formats.
 
-        Args:
-            cloud_endpoint: Cloud endpoint URL or hostname
-
-        Returns:
-            Parsed hostname with .grpc. inserted if needed
-
-        Raises:
-            ValueError: If endpoint cannot be parsed
+        .. deprecated:: 25.1.0
+            Dgraph Cloud service has been discontinued. This method will be
+            removed in version 26.0.0. Use the standard AsyncDgraphClientStub
+            constructor with grpc.ssl_channel_credentials() instead.
         """
-        # Normalize to have scheme for consistent parsing
+        warnings.warn(
+            "parse_host() is deprecated as Dgraph Cloud has been discontinued. "
+            "This method will be removed in version 26.0.0. "
+            "Use the standard AsyncDgraphClientStub constructor with SSL credentials instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+
         endpoint = cloud_endpoint
         if "://" not in endpoint:
             endpoint = f"//{endpoint}"
 
-        # Parse URL to extract hostname (handles IPv6, ports, etc.)
         try:
             parsed = urlparse(endpoint)
             host = parsed.hostname if parsed.hostname else cloud_endpoint
         except Exception:
-            # Fallback for malformed URLs
             host = cloud_endpoint
 
-        # Remove any port that wasn't caught by hostname parsing
-        if ":" in host and "[" not in host:  # Not IPv6
+        if ":" in host and "[" not in host:
             host = host.split(":", 1)[0]
 
-        # Only insert .grpc. if:
-        # 1. Not already present
-        # 2. Host has at least 2 labels (e.g., "example.com" but not "localhost")
-        # 3. Not an IP address
         if ".grpc." not in host and "." in host:
-            # Check it's not an IPv4 address
             labels = host.split(".")
             if len(labels) >= 2 and not all(label.isdigit() for label in labels):
-                host = f'{labels[0]}.grpc.{".".join(labels[1:])}'
+                host = f"{labels[0]}.grpc.{'.'.join(labels[1:])}"
 
         return host
 
     @staticmethod
-    def from_cloud(cloud_endpoint, api_key, options=None):
+    def from_cloud(
+        cloud_endpoint: str,
+        api_key: str,
+        options: list[tuple[str, Any]] | None = None,
+    ) -> AsyncDgraphClientStub:
         """Returns async Dgraph Client stub for Dgraph Cloud endpoint.
 
-        Usage:
-            import pydgraph
-            client_stub = pydgraph.AsyncDgraphClientStub.from_cloud("cloud_endpoint", "api-key")
-            client = pydgraph.AsyncDgraphClient(client_stub)
+        .. deprecated:: 25.1.0
+            Dgraph Cloud service has been discontinued. This method will be
+            removed in version 26.0.0. Use the standard AsyncDgraphClientStub
+            constructor with grpc.ssl_channel_credentials() instead.
 
-        Args:
-            cloud_endpoint: Dgraph Cloud endpoint (can be grpc or graphql endpoint)
-            api_key: API key for authentication
-            options: gRPC channel options
+        Example migration:
+            Old: stub = AsyncDgraphClientStub.from_cloud(endpoint, api_key)
 
-        Returns:
-            AsyncDgraphClientStub instance configured for Dgraph Cloud
+            New:
+                import grpc
+                creds = grpc.ssl_channel_credentials()
+                call_creds = grpc.metadata_call_credentials(
+                    lambda _c, cb: cb((("authorization", api_key),), None)
+                )
+                composite = grpc.composite_channel_credentials(creds, call_creds)
+                stub = AsyncDgraphClientStub(endpoint, composite)
         """
+        warnings.warn(
+            "from_cloud() is deprecated as Dgraph Cloud has been discontinued. "
+            "This method will be removed in version 26.0.0. "
+            "Use the standard AsyncDgraphClientStub constructor with SSL credentials. "
+            "See the docstring for migration examples.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+
         host = AsyncDgraphClientStub.parse_host(cloud_endpoint)
         creds = grpc.ssl_channel_credentials()
         call_credentials = grpc.metadata_call_credentials(
-            lambda context, callback: callback((("authorization", api_key),), None)
+            lambda _context, callback: callback((("authorization", api_key),), None)
         )
         composite_credentials = grpc.composite_channel_credentials(
             creds, call_credentials
         )
 
-        # Create new list to avoid mutating caller's options
         opts = list(options) if options is not None else []
-
-        # Add http proxy setting if not already present
         if not any(k == "grpc.enable_http_proxy" for k, _ in opts):
             opts.append(("grpc.enable_http_proxy", 0))
 
-        client_stub = AsyncDgraphClientStub(
-            "{host}:{port}".format(host=host, port="443"),
+        return AsyncDgraphClientStub(
+            f"{host}:443",
             composite_credentials,
             options=opts,
         )
-        return client_stub
+
+    async def close(self) -> None:
+        """Close the async channel gracefully."""
+        with contextlib.suppress(Exception):
+            await self.channel.close()
