@@ -1,11 +1,14 @@
-# SPDX-FileCopyrightText: © 2017-2025 Istari Digital, Inc.
+# SPDX-FileCopyrightText: © 2017-2026 Istari Digital, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
 """Dgraph python client."""
+from __future__ import annotations
 
 import contextlib
-import random
+import secrets
 import urllib.parse
+from collections.abc import Iterator
+from typing import TYPE_CHECKING, Any
 
 import grpc
 
@@ -14,28 +17,36 @@ from pydgraph.client_stub import DgraphClientStub
 from pydgraph.meta import VERSION
 from pydgraph.proto import api_pb2 as api
 
+if TYPE_CHECKING:
+    from pydgraph.txn import Txn
+
 __author__ = "Mohit Ranka <mohitranka@gmail.com>"
 __maintainer__ = "Istari Digital, Inc. <dgraph-admin@istaridigital.com>"
 __version__ = VERSION
 __status__ = "development"
 
 
-class DgraphClient(object):
+class DgraphClient:
     """Creates a new Client for interacting with Dgraph.
 
     The client can be backed by multiple connections (to the same server, or
     multiple servers in a cluster).
     """
 
-    def __init__(self, *clients):
+    def __init__(self, *clients: DgraphClientStub) -> None:
         if not clients:
             raise ValueError("No clients provided in DgraphClient constructor")
 
         self._clients = clients[:]
         self._jwt = api.Jwt()
-        self._login_metadata = []
+        self._login_metadata: list[tuple[str, str]] = []
 
-    def check_version(self, timeout=None, metadata=None, credentials=None):
+    def check_version(
+        self,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> str:
         """Returns the version of Dgraph if the server is ready to accept requests."""
 
         new_metadata = self.add_login_metadata(metadata)
@@ -48,7 +59,6 @@ class DgraphClient(object):
                 metadata=new_metadata,
                 credentials=credentials,
             )
-            return response.tag
         except Exception as error:
             if util.is_jwt_expired(error):
                 self.retry_login()
@@ -59,11 +69,19 @@ class DgraphClient(object):
                     metadata=new_metadata,
                     credentials=credentials,
                 )
-                return response.tag
             else:
-                raise error
+                raise
 
-    def login(self, userid, password, timeout=None, metadata=None, credentials=None):
+        return response.tag
+
+    def login(
+        self,
+        userid: str,
+        password: str,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> None:
         """Attempts a login via this client."""
 
         return self.login_into_namespace(
@@ -76,8 +94,14 @@ class DgraphClient(object):
         )
 
     def login_into_namespace(
-        self, userid, password, namespace, timeout=None, metadata=None, credentials=None
-    ):
+        self,
+        userid: str,
+        password: str,
+        namespace: int,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> None:
         """Attempts a login into a namespace via this client."""
 
         login_req = api.LoginRequest()
@@ -92,7 +116,12 @@ class DgraphClient(object):
         self._jwt.ParseFromString(response.json)
         self._login_metadata = [("accessjwt", self._jwt.access_jwt)]
 
-    def retry_login(self, timeout=None, metadata=None, credentials=None):
+    def retry_login(
+        self,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> None:
         """Attempts a retry login via this client."""
 
         if len(self._jwt.refresh_jwt) == 0:
@@ -108,7 +137,13 @@ class DgraphClient(object):
         self._jwt.ParseFromString(response.json)
         self._login_metadata = [("accessjwt", self._jwt.access_jwt)]
 
-    def alter(self, operation, timeout=None, metadata=None, credentials=None):
+    def alter(
+        self,
+        operation: api.Operation,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Payload:
         """Runs a modification via this client."""
 
         new_metadata = self.add_login_metadata(metadata)
@@ -131,13 +166,15 @@ class DgraphClient(object):
                         metadata=new_metadata,
                         credentials=credentials,
                     )
-                except Exception as error:
-                    self._common_except_alter(error)
+                except Exception as retry_error:
+                    self._common_except_alter(retry_error)
+                    raise  # This should never be reached due to _common_except_alter raising
             else:
                 self._common_except_alter(error)
+                raise  # This should never be reached due to _common_except_alter raising
 
     @staticmethod
-    def _common_except_alter(error):
+    def _common_except_alter(error: Exception) -> None:
         if util.is_retriable_error(error):
             raise errors.RetriableError(error)
 
@@ -146,7 +183,13 @@ class DgraphClient(object):
 
         raise error
 
-    def async_alter(self, operation, timeout=None, metadata=None, credentials=None):
+    def async_alter(
+        self,
+        operation: api.Operation,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> grpc.Future:
         """The async version of alter."""
 
         new_metadata = self.add_login_metadata(metadata)
@@ -155,13 +198,19 @@ class DgraphClient(object):
         )
 
     @staticmethod
-    def handle_alter_future(future):
+    def handle_alter_future(future: grpc.Future) -> api.Payload:
         try:
             return future.result()
         except Exception as error:
             DgraphClient._common_except_alter(error)
+            raise  # This should never be reached due to _common_except_alter raising
 
-    def drop_all(self, timeout=None, metadata=None, credentials=None):
+    def drop_all(
+        self,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Payload:
         """Drops all data and schema from the Dgraph instance.
 
         Args:
@@ -180,7 +229,12 @@ class DgraphClient(object):
             operation, timeout=timeout, metadata=metadata, credentials=credentials
         )
 
-    def drop_data(self, timeout=None, metadata=None, credentials=None):
+    def drop_data(
+        self,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Payload:
         """Drops all data from the Dgraph instance while preserving the schema.
 
         Args:
@@ -194,12 +248,18 @@ class DgraphClient(object):
         Raises:
             Exception: If the request fails.
         """
-        operation = api.Operation(drop_op="DATA")
+        operation = api.Operation(drop_op=api.Operation.DropOp.DATA)
         return self.alter(
             operation, timeout=timeout, metadata=metadata, credentials=credentials
         )
 
-    def drop_predicate(self, predicate, timeout=None, metadata=None, credentials=None):
+    def drop_predicate(
+        self,
+        predicate: str,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Payload:
         """Drops a predicate and its associated data from the Dgraph instance.
 
         Args:
@@ -216,12 +276,20 @@ class DgraphClient(object):
         """
         if not predicate:
             raise ValueError("predicate cannot be empty")
-        operation = api.Operation(drop_op="ATTR", drop_value=predicate)
+        operation = api.Operation(
+            drop_op=api.Operation.DropOp.ATTR, drop_value=predicate
+        )
         return self.alter(
             operation, timeout=timeout, metadata=metadata, credentials=credentials
         )
 
-    def drop_type(self, type_name, timeout=None, metadata=None, credentials=None):
+    def drop_type(
+        self,
+        type_name: str,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Payload:
         """Drops a type definition from the DQL schema.
 
         Note: This only removes the type definition from the schema. No data is
@@ -242,12 +310,20 @@ class DgraphClient(object):
         """
         if not type_name:
             raise ValueError("type_name cannot be empty")
-        operation = api.Operation(drop_op="TYPE", drop_value=type_name)
+        operation = api.Operation(
+            drop_op=api.Operation.DropOp.TYPE, drop_value=type_name
+        )
         return self.alter(
             operation, timeout=timeout, metadata=metadata, credentials=credentials
         )
 
-    def set_schema(self, schema, timeout=None, metadata=None, credentials=None):
+    def set_schema(
+        self,
+        schema: str,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Payload:
         """Sets the DQL schema for the Dgraph instance.
 
         Args:
@@ -269,7 +345,9 @@ class DgraphClient(object):
             operation, timeout=timeout, metadata=metadata, credentials=credentials
         )
 
-    def txn(self, read_only=False, best_effort=False, **commit_kwargs):
+    def txn(
+        self, read_only: bool = False, best_effort: bool = False, **commit_kwargs: Any
+    ) -> Txn:
         """Creates a transaction."""
         return txn.Txn(
             self, read_only=read_only, best_effort=best_effort, **commit_kwargs
@@ -277,15 +355,15 @@ class DgraphClient(object):
 
     def run_dql(
         self,
-        dql_query,
-        vars=None,
-        read_only=False,
-        best_effort=False,
-        resp_format="JSON",
-        timeout=None,
-        metadata=None,
-        credentials=None,
-    ):
+        dql_query: str,
+        vars: dict[str, str] | None = None,  # noqa: A002
+        read_only: bool = False,
+        best_effort: bool = False,
+        resp_format: str = "JSON",
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Response:
         """
         Runs a DQL query or mutation via this client.
 
@@ -327,21 +405,24 @@ class DgraphClient(object):
                 new_metadata.append(("namespace", "0"))
 
         # Convert string format to enum if needed
+        format_value: int
         if isinstance(resp_format, str):
             if resp_format.upper() == "JSON":
-                resp_format = api.Request.RespFormat.JSON
+                format_value = api.Request.RespFormat.JSON
             elif resp_format.upper() == "RDF":
-                resp_format = api.Request.RespFormat.RDF
+                format_value = api.Request.RespFormat.RDF
             else:
                 raise ValueError(
                     f"Invalid resp_format: {resp_format}. Must be 'JSON' or 'RDF'"
                 )
+        else:
+            format_value = resp_format
         req = api.RunDQLRequest(
             dql_query=dql_query,
             vars=vars,
             read_only=read_only,
             best_effort=best_effort,
-            resp_format=resp_format,
+            resp_format=format_value,
         )
         try:
             return self.any_client().run_dql(
@@ -360,20 +441,19 @@ class DgraphClient(object):
                     metadata=new_metadata,
                     credentials=credentials,
                 )
-            else:
-                raise error
+            raise
 
     def run_dql_with_vars(
         self,
-        dql_query,
-        vars,
-        read_only=False,
-        best_effort=False,
-        resp_format="JSON",
-        timeout=None,
-        metadata=None,
-        credentials=None,
-    ):
+        dql_query: str,
+        vars: dict[str, str],  # noqa: A002
+        read_only: bool = False,
+        best_effort: bool = False,
+        resp_format: str = "JSON",
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> api.Response:
         """
         Runs a DQL query or mutation with variables via this client.
 
@@ -408,7 +488,13 @@ class DgraphClient(object):
             credentials=credentials,
         )
 
-    def allocate_uids(self, how_many, timeout=None, metadata=None, credentials=None):
+    def allocate_uids(
+        self,
+        how_many: int,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> tuple[int, int]:
         """
         AllocateUIDs allocates a given number of Node UIDs in the Graph and returns a start and end UIDs,
         end excluded. The UIDs in the range [start, end) can then be used by the client in the mutations
@@ -430,8 +516,12 @@ class DgraphClient(object):
         return self._allocate_ids(how_many, api.UID, timeout, metadata, credentials)
 
     def allocate_timestamps(
-        self, how_many, timeout=None, metadata=None, credentials=None
-    ):
+        self,
+        how_many: int,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> tuple[int, int]:
         """
         AllocateTimestamps gets a sequence of timestamps allocated from Dgraph. These timestamps can be
         used in bulk loader and similar applications.
@@ -450,8 +540,12 @@ class DgraphClient(object):
         return self._allocate_ids(how_many, api.TS, timeout, metadata, credentials)
 
     def allocate_namespaces(
-        self, how_many, timeout=None, metadata=None, credentials=None
-    ):
+        self,
+        how_many: int,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> tuple[int, int]:
         """
         AllocateNamespaces allocates a given number of namespaces in the Graph and returns a start and end
         namespaces, end excluded. The namespaces in the range [start, end) can then be used by the client.
@@ -472,8 +566,13 @@ class DgraphClient(object):
         return self._allocate_ids(how_many, api.NS, timeout, metadata, credentials)
 
     def _allocate_ids(
-        self, how_many, lease_type, timeout=None, metadata=None, credentials=None
-    ):
+        self,
+        how_many: int,
+        lease_type: int,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> tuple[int, int]:
         """
         Helper method to allocate IDs of different types (UIDs, timestamps, namespaces).
 
@@ -510,10 +609,14 @@ class DgraphClient(object):
                     credentials=credentials,
                 )
                 return response.start, response.end + 1
-            else:
-                raise error
+            raise
 
-    def create_namespace(self, timeout=None, metadata=None, credentials=None):
+    def create_namespace(
+        self,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> int:
         """Creates a new namespace and returns its ID.
 
         Args:
@@ -538,7 +641,6 @@ class DgraphClient(object):
                 metadata=new_metadata,
                 credentials=credentials,
             )
-            return response.namespace
         except Exception as error:
             if util.is_jwt_expired(error):
                 self.retry_login()
@@ -549,11 +651,18 @@ class DgraphClient(object):
                     metadata=new_metadata,
                     credentials=credentials,
                 )
-                return response.namespace
             else:
-                raise error
+                raise
 
-    def drop_namespace(self, namespace, timeout=None, metadata=None, credentials=None):
+        return response.namespace
+
+    def drop_namespace(
+        self,
+        namespace: int,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> Any:
         """Drops the specified namespace. If the namespace does not exist, the request will still
         succeed.
 
@@ -578,7 +687,6 @@ class DgraphClient(object):
                 metadata=new_metadata,
                 credentials=credentials,
             )
-            return response
         except Exception as error:
             if util.is_jwt_expired(error):
                 self.retry_login()
@@ -589,11 +697,17 @@ class DgraphClient(object):
                     metadata=new_metadata,
                     credentials=credentials,
                 )
-                return response
             else:
-                raise error
+                raise
 
-    def list_namespaces(self, timeout=None, metadata=None, credentials=None):
+        return response
+
+    def list_namespaces(
+        self,
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> dict[int, Any]:
         """Lists all namespaces.
 
         Args:
@@ -630,22 +744,23 @@ class DgraphClient(object):
                     credentials=credentials,
                 )
                 return dict(response.namespaces)
-            else:
-                raise error
+            raise
 
-    def any_client(self):
+    def any_client(self) -> DgraphClientStub:
         """Returns a random gRPC client so that requests are distributed evenly among them."""
 
-        return random.choice(self._clients)  # nosec # pylint: disable=insecure-random
+        return secrets.choice(self._clients)
 
-    def add_login_metadata(self, metadata):
+    def add_login_metadata(
+        self, metadata: list[tuple[str, str]] | None
+    ) -> list[tuple[str, str]]:
         new_metadata = list(self._login_metadata)
         if not metadata:
             return new_metadata
         new_metadata.extend(metadata)
         return new_metadata
 
-    def close(self):
+    def close(self) -> None:
         for client in self._clients:
             client.close()
 
@@ -654,10 +769,10 @@ class DgraphClient(object):
         self,
         read_only: bool = False,
         best_effort: bool = False,
-        timeout=None,
-        metadata=None,
-        credentials=None,
-    ):
+        timeout: float | None = None,
+        metadata: list[tuple[str, str]] | None = None,
+        credentials: grpc.CallCredentials | None = None,
+    ) -> Iterator[Txn]:
         """Start a managed transaction.
 
         Note
@@ -669,13 +784,11 @@ class DgraphClient(object):
             yield tx
             if not read_only and not tx._finished:
                 tx.commit(timeout=timeout, metadata=metadata, credentials=credentials)
-        except Exception as e:
-            raise e
         finally:
             tx.discard()
 
 
-def open(connection_string: str) -> DgraphClient:
+def open(connection_string: str) -> DgraphClient:  # noqa: A001, C901
     """Open a new Dgraph client. Use client.close() to close the client.
 
     Args:
@@ -690,14 +803,15 @@ def open(connection_string: str) -> DgraphClient:
 
     try:
         parsed_url = urllib.parse.urlparse(connection_string)
-        if not parsed_url.scheme == "dgraph":
-            raise ValueError("Invalid connection string: scheme must be 'dgraph'")
-        if not parsed_url.hostname:
-            raise ValueError("Invalid connection string: hostname required")
-        if not parsed_url.port:
-            raise ValueError("Invalid connection string: port required")
     except Exception as e:
         raise ValueError(f"Failed to parse connection string: {e}") from e
+
+    if not parsed_url.scheme == "dgraph":
+        raise ValueError("Invalid connection string: scheme must be 'dgraph'")
+    if not parsed_url.hostname:
+        raise ValueError("Invalid connection string: hostname required")
+    if not parsed_url.port:
+        raise ValueError("Invalid connection string: port required")
 
     host = parsed_url.hostname
     port = parsed_url.port
@@ -738,21 +852,20 @@ def open(connection_string: str) -> DgraphClient:
     if "namespace" in params:
         try:
             namespace = int(params["namespace"])
-            if namespace < 0:
-                raise ValueError(f"namespace must be >= 0, got {namespace}")
         except ValueError as e:
-            if "namespace must be >= 0" in str(e):
-                raise e
             raise TypeError(
                 f"namespace must be an integer, got '{params['namespace']}'"
             ) from e
+
+        if namespace < 0:
+            raise ValueError(f"namespace must be >= 0, got {namespace}")
         if not username:
             raise ValueError("username/password required when namespace is provided")
 
     if auth_header:
         options = [("grpc.enable_http_proxy", 0)]
         call_credentials = grpc.metadata_call_credentials(
-            lambda context, callback: callback((("authorization", auth_header),), None)
+            lambda _context, callback: callback((("authorization", auth_header),), None)
         )
         credentials = grpc.composite_channel_credentials(
             grpc.ssl_channel_credentials(), call_credentials
@@ -763,6 +876,8 @@ def open(connection_string: str) -> DgraphClient:
 
     if username:
         thirty_seconds = 30
+        if password is None:
+            raise ValueError("password is required when username is provided")
         client.login_into_namespace(
             username, password, namespace, timeout=thirty_seconds
         )
