@@ -1,7 +1,11 @@
-# SPDX-FileCopyrightText: © 2017-2025 Istari Digital, Inc.
+# SPDX-FileCopyrightText: © 2017-2026 Istari Digital, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
 """Tests to verify upsert directive."""
+
+from __future__ import annotations
+
+from collections.abc import Callable
 
 __author__ = "Shailesh Kochhar <shailesh.kochhar@gmail.com>"
 __maintainer__ = "Istari Digital, Inc. <dgraph-admin@istaridigital.com>"
@@ -12,6 +16,7 @@ import multiprocessing
 import multiprocessing.dummy as mpd
 import time
 import unittest
+from typing import Any
 
 import pydgraph
 from tests import helper
@@ -25,8 +30,8 @@ AGES = [20, 25, 30, 35]
 class TestAccountUpsert(helper.ClientIntegrationTestCase):
     """Tests to verify upsert directive."""
 
-    def setUp(self):
-        super(TestAccountUpsert, self).setUp()
+    def setUp(self) -> None:
+        super().setUp()
 
         self.accounts = [
             {"first": f, "last": ln, "age": a}
@@ -47,24 +52,29 @@ class TestAccountUpsert(helper.ClientIntegrationTestCase):
         """,
         )
 
-    def test_account_upsert(self):
+    def test_account_upsert(self) -> None:
         """Run upserts concurrently."""
         self.do_upserts(self.accounts, CONCURRENCY, upsert_account)
         self.assert_changes(FIRSTS, self.accounts)
 
-    def test_account_upsert_block(self):
+    def test_account_upsert_block(self) -> None:
         """Run upserts concurrently using upsert block."""
         self.do_upserts(self.accounts, CONCURRENCY, upsert_account_upsert_block)
         self.assert_changes(FIRSTS, self.accounts)
 
-    def do_upserts(self, account_list, concurrency, upsert_func):
+    def do_upserts(
+        self,
+        account_list: list[dict[str, object]],
+        concurrency: int,
+        upsert_func: Callable[..., None],
+    ) -> None:
         """Runs the upsert command for the accounts in `account_list`. Execution
         happens in concurrent processes."""
 
         success_ctr = multiprocessing.Value("i", 0, lock=True)
         retry_ctr = multiprocessing.Value("i", 0, lock=True)
 
-        def _updater(acct):
+        def _updater(acct: dict[str, object]) -> None:
             upsert_func(
                 addr=self.TEST_SERVER_ADDR,
                 account=acct,
@@ -82,7 +92,9 @@ class TestAccountUpsert(helper.ClientIntegrationTestCase):
         _ = [res.get() for res in results]
         pool.close()
 
-    def assert_changes(self, firsts, accounts):
+    def assert_changes(
+        self, firsts: list[str], accounts: list[dict[str, object]]
+    ) -> None:
         """Will check to see changes have been made."""
 
         query = """{{
@@ -91,25 +103,25 @@ class TestAccountUpsert(helper.ClientIntegrationTestCase):
                 last
                 age
             }}
-        }}""".format(
-            " ".join(firsts)
-        )
+        }}""".format(" ".join(firsts))
         logging.debug(query)
         result = json.loads(self.client.txn(read_only=True).query(query).json)
 
         account_set = set()
         for acct in result["all"]:
-            self.assertTrue(acct["first"] is not None)
-            self.assertTrue(acct["last"] is not None)
-            self.assertTrue(acct["age"] is not None)
+            assert acct["first"] is not None
+            assert acct["last"] is not None
+            assert acct["age"] is not None
             account_set.add("{first}_{last}_{age}".format(**acct))
 
-        self.assertEqual(len(account_set), len(accounts))
+        assert len(account_set) == len(accounts)
         for acct in accounts:
-            self.assertTrue("{first}_{last}_{age}".format(**acct) in account_set)
+            assert "{first}_{last}_{age}".format(**acct) in account_set
 
 
-def upsert_account(addr, account, success_ctr, retry_ctr):
+def upsert_account(
+    addr: str, account: dict[str, object], success_ctr: Any, retry_ctr: Any
+) -> None:
     """Runs upsert operation."""
     client = helper.create_client(addr)
     client.login("groot", "password")
@@ -117,9 +129,7 @@ def upsert_account(addr, account, success_ctr, retry_ctr):
         acct(func:eq(first, "{first}")) @filter(eq(last, "{last}") AND eq(age, {age})) {{
             uid
         }}
-    }}""".format(
-        **account
-    )
+    }}""".format(**account)
 
     last_update_time = time.time() - 10000
     while True:
@@ -131,7 +141,7 @@ def upsert_account(addr, account, success_ctr, retry_ctr):
         try:
             result = json.loads(txn.query(query).json)
             assert len(result["acct"]) <= 1, (
-                "Lookup of account %s found " "multiple accounts" % account
+                f"Lookup of account {account} found multiple accounts"
             )
 
             if not result["acct"]:
@@ -140,38 +150,37 @@ def upsert_account(addr, account, success_ctr, retry_ctr):
                     _:acct <first> "{first}" .
                     _:acct <last> "{last}" .
                     _:acct <age>  "{age}"^^<xs:int> .
-                """.format(
-                    **account
-                )
+                """.format(**account)
                 created = txn.mutate(set_nquads=nquads)
                 uid = created.uids.get("acct")
-                assert uid is not None and uid != "", "Account with uid None"
+                assert uid is not None, "Account with uid None"
+                assert uid != "", "Account with empty uid"
             else:
                 # account exists, read the uid
                 acct = result["acct"][0]
                 uid = acct["uid"]
                 assert uid is not None, "Account with uid None"
 
-            updatequads = '<{0}> <when> "{1:d}"^^<xs:int> .'.format(
-                uid, int(time.time())
-            )
+            updatequads = f'<{uid}> <when> "{int(time.time()):d}"^^<xs:int> .'
             txn.mutate(set_nquads=updatequads)
             txn.commit()
 
             with success_ctr.get_lock():
                 success_ctr.value += 1
-
-            # txn successful, break the loop
-            return
         except pydgraph.AbortedError:
             with retry_ctr.get_lock():
                 retry_ctr.value += 1
             # txn failed, retry the loop
+        else:
+            # txn successful, break the loop
+            return
         finally:
             txn.discard()
 
 
-def upsert_account_upsert_block(addr, account, success_ctr, retry_ctr):
+def upsert_account_upsert_block(
+    addr: str, account: dict[str, object], success_ctr: Any, retry_ctr: Any
+) -> None:
     """Runs upsert operation."""
     client = helper.create_client(addr)
     client.login("groot", "password")
@@ -179,9 +188,7 @@ def upsert_account_upsert_block(addr, account, success_ctr, retry_ctr):
         acct(func:eq(first, "{first}")) @filter(eq(last, "{last}") AND eq(age, {age})) {{
             u as uid
         }}
-    }}""".format(
-        **account
-    )
+    }}""".format(**account)
 
     last_update_time = time.time() - 10000
     while True:
@@ -195,16 +202,14 @@ def upsert_account_upsert_block(addr, account, success_ctr, retry_ctr):
                 uid(u) <first> "{first}" .
                 uid(u) <last> "{last}" .
                 uid(u) <age>  "{age}"^^<xs:int> .
-            """.format(
-                **account
-            )
+            """.format(**account)
             mutation = txn.create_mutation(set_nquads=nquads)
             request = txn.create_request(
                 query=query, mutations=[mutation], commit_now=True
             )
             txn.do_request(request)
 
-            updatequads = 'uid(u) <when> "{0:d}"^^<xs:int> .'.format(int(time.time()))
+            updatequads = f'uid(u) <when> "{int(time.time()):d}"^^<xs:int> .'
             txn = client.txn()
             mutation = txn.create_mutation(set_nquads=updatequads)
             request = txn.create_request(
@@ -214,18 +219,18 @@ def upsert_account_upsert_block(addr, account, success_ctr, retry_ctr):
 
             with success_ctr.get_lock():
                 success_ctr.value += 1
-
-            # txn successful, break the loop
-            return
         except pydgraph.AbortedError:
             with retry_ctr.get_lock():
                 retry_ctr.value += 1
             # txn failed, retry the loop
+        else:
+            # txn successful, break the loop
+            return
         finally:
             txn.discard()
 
 
-def suite():
+def suite() -> unittest.TestSuite:
     """Returns a test suite object."""
     suite_obj = unittest.TestSuite()
     suite_obj.addTest(TestAccountUpsert())
