@@ -1,7 +1,9 @@
-# SPDX-FileCopyrightText: © 2017-2025 Istari Digital, Inc.
+# SPDX-FileCopyrightText: © 2017-2026 Istari Digital, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
 """Tests to verify async client methods."""
+
+from __future__ import annotations
 
 __author__ = "Martin Martinez Rivera"
 __maintainer__ = "Istari Digital, Inc. <dgraph-admin@istaridigital.com>"
@@ -15,11 +17,11 @@ from tests import helper
 class TestAsync(helper.ClientIntegrationTestCase):
     server_addr = "localhost:9180"
 
-    def setUp(self):
-        super(TestAsync, self).setUp()
+    def setUp(self) -> None:
+        super().setUp()
         helper.drop_all(self.client)
 
-    def test_mutation_and_query(self):
+    def test_mutation_and_query(self) -> None:
         """Runs mutation and queries asyncronously."""
         alter_future = self.client.async_alter(
             pydgraph.Operation(schema="name: string @index(term) .")
@@ -51,7 +53,6 @@ class TestAsync(helper.ClientIntegrationTestCase):
         txn = self.client.txn()
         query_future = txn.async_query(query, variables={"$a": "Alice"})
         response = pydgraph.Txn.handle_query_future(query_future)
-        self.assertEqual(
-            [{"name": "Alice", "follows": [{"name": "Greg"}]}],
-            json.loads(response.json).get("me"),
-        )
+        assert json.loads(response.json).get("me") == [
+            {"name": "Alice", "follows": [{"name": "Greg"}]}
+        ]
