@@ -1,12 +1,17 @@
-# SPDX-FileCopyrightText: © 2017-2025 Istari Digital, Inc.
+# SPDX-FileCopyrightText: © 2017-2026 Istari Digital, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
 """Tests client stub."""
+
+from __future__ import annotations
 
 __author__ = "Garvit Pahal"
 __maintainer__ = "Istari Digital, Inc. <dgraph-admin@istaridigital.com>"
 
 import unittest
+from typing import Any
+
+import pytest
 
 import pydgraph
 
@@ -16,90 +21,72 @@ from . import helper
 class TestDgraphClientStub(helper.ClientIntegrationTestCase):
     """Tests client stub."""
 
-    def validate_version_object(self, version):
+    def validate_version_object(self, version: Any) -> None:
         tag = version.tag
-        self.assertIsInstance(tag, str)
+        assert isinstance(tag, str)
 
-    def check_version(self, stub):
+    def check_version(self, stub: Any) -> None:
         self.validate_version_object(stub.check_version(pydgraph.Check()))
 
-    def test_constructor(self):
+    def test_constructor(self) -> None:
         self.check_version(pydgraph.DgraphClientStub(addr=self.TEST_SERVER_ADDR))
 
-    def test_timeout(self):
-        with self.assertRaises(Exception):  # noqa: B017
+    def test_timeout(self) -> None:
+        with pytest.raises(Exception):
             pydgraph.DgraphClientStub(self.TEST_SERVER_ADDR).check_version(
                 pydgraph.Check(), timeout=-1
             )
 
-    def test_close(self):
+    def test_close(self) -> None:
         client_stub = pydgraph.DgraphClientStub(addr=self.TEST_SERVER_ADDR)
         self.check_version(client_stub)
         client_stub.close()
-        with self.assertRaises(Exception):  # noqa: B017
+        with pytest.raises(Exception):
             client_stub.check_version(pydgraph.Check())
 
 
-class TestFromCloud(unittest.TestCase):
-    """Tests the from_cloud function"""
-
-    def test_from_cloud(self):
-        testcases = [
-            {"endpoint": "godly.grpc.region.aws.cloud.dgraph.io"},
-            {"endpoint": "godly.grpc.region.aws.cloud.dgraph.io:443"},
-            {"endpoint": "https://godly.grpc.region.aws.cloud.dgraph.io:443"},
-            {"endpoint": "https://godly.region.aws.cloud.dgraph.io/graphql"},
-            {"endpoint": "godly.region.aws.cloud.dgraph.io"},
-            {"endpoint": "https://godly.region.aws.cloud.dgraph.io"},
-            {"endpoint": "godly.region.aws.cloud.dgraph.io:random"},
-            {"endpoint": "random:url", "error": True},
-            {"endpoint": "google", "error": True},
-        ]
-
-        for case in testcases:
-            try:
-                pydgraph.DgraphClientStub.from_cloud(case["endpoint"], "api-key")
-            except IndexError as e:
-                if not case["error"]:
-                    # we didn't expect an error
-                    raise (e)
-
-
 class TestDgraphClientStubContextManager(helper.ClientIntegrationTestCase):
-    def setUp(self):
-        super(TestDgraphClientStubContextManager, self).setUp()
+    def setUp(self) -> None:
+        super().setUp()
 
-    def test_context_manager(self):
+    def check_version(self, stub: Any) -> None:
+        """Helper method to check version using the stub."""
+        version = stub.check_version(pydgraph.Check())
+        assert version is not None
+
+    def test_context_manager(self) -> None:
         """Test basic context manager usage for DgraphClientStub."""
         with pydgraph.DgraphClientStub(addr=self.TEST_SERVER_ADDR) as client_stub:
             ver = client_stub.check_version(pydgraph.Check())
-            self.assertIsNotNone(ver)
+            assert ver is not None
 
-    def test_context_manager_code_exception(self):
+    def test_context_manager_code_exception(self) -> None:
         """Test that exceptions within context manager are properly handled."""
-        with self.assertRaises(AttributeError):
-            with pydgraph.DgraphClientStub(addr=self.TEST_SERVER_ADDR) as client_stub:
-                self.check_version(client_stub)  # AttributeError: no such method
+        with (
+            pytest.raises(AttributeError),
+            pydgraph.DgraphClientStub(addr=self.TEST_SERVER_ADDR) as client_stub,
+        ):
+            client_stub.nonexistent_method()  # type: ignore[attr-defined]  # Intentionally calling non-existent method
 
-    def test_context_manager_function_wrapper(self):
+    def test_context_manager_function_wrapper(self) -> None:
         """Test the client_stub() function wrapper for context manager."""
         with pydgraph.client_stub(addr=self.TEST_SERVER_ADDR) as client_stub:
             ver = client_stub.check_version(pydgraph.Check())
-            self.assertIsNotNone(ver)
+            assert ver is not None
 
-    def test_context_manager_closes_stub(self):
+    def test_context_manager_closes_stub(self) -> None:
         """Test that the stub is properly closed after exiting context manager."""
         stub = None
         with pydgraph.DgraphClientStub(addr=self.TEST_SERVER_ADDR) as client_stub:
             stub = client_stub
             ver = client_stub.check_version(pydgraph.Check())
-            self.assertIsNotNone(ver)
+            assert ver is not None
 
         # After exiting context, stub should be closed and unusable
-        with self.assertRaises(Exception):  # noqa: B017
+        with pytest.raises(Exception):
             stub.check_version(pydgraph.Check())
 
-    def test_context_manager_with_client(self):
+    def test_context_manager_with_client(self) -> None:
         """Test using DgraphClientStub context manager with DgraphClient."""
         with pydgraph.DgraphClientStub(addr=self.TEST_SERVER_ADDR) as client_stub:
             client = pydgraph.DgraphClient(client_stub)
@@ -109,43 +96,44 @@ class TestDgraphClientStubContextManager(helper.ClientIntegrationTestCase):
             txn = client.txn(read_only=True)
             query = "{ me(func: has(name)) { name } }"
             resp = txn.query(query)
-            self.assertIsNotNone(resp)
+            assert resp is not None
 
-    def test_context_manager_exception_still_closes(self):
+    def test_context_manager_exception_still_closes(self) -> None:
         """Test that stub is closed even when an exception occurs."""
         stub_ref = None
         try:
             with pydgraph.DgraphClientStub(addr=self.TEST_SERVER_ADDR) as client_stub:
                 stub_ref = client_stub
                 client_stub.check_version(pydgraph.Check())
-                raise ValueError("Test exception")
+                raise ValueError("Test exception")  # noqa: TRY301
         except ValueError:
             pass
 
         # Stub should still be closed despite the exception
-        with self.assertRaises(Exception):  # noqa: B017
+        assert stub_ref is not None
+        with pytest.raises(Exception):
             stub_ref.check_version(pydgraph.Check())
 
-    def test_context_manager_function_wrapper_closes(self):
+    def test_context_manager_function_wrapper_closes(self) -> None:
         """Test that client_stub() function wrapper properly closes the stub."""
         stub_ref = None
         with pydgraph.client_stub(addr=self.TEST_SERVER_ADDR) as client_stub:
             stub_ref = client_stub
             ver = client_stub.check_version(pydgraph.Check())
-            self.assertIsNotNone(ver)
+            assert ver is not None
 
         # Stub should be closed after exiting
-        with self.assertRaises(Exception):  # noqa: B017
+        with pytest.raises(Exception):
             stub_ref.check_version(pydgraph.Check())
 
-    def test_context_manager_multiple_operations(self):
+    def test_context_manager_multiple_operations(self) -> None:
         """Test performing multiple operations within context manager."""
         with pydgraph.DgraphClientStub(addr=self.TEST_SERVER_ADDR) as client_stub:
             # Check version multiple times
             ver1 = client_stub.check_version(pydgraph.Check())
             ver2 = client_stub.check_version(pydgraph.Check())
-            self.assertIsNotNone(ver1)
-            self.assertIsNotNone(ver2)
+            assert ver1 is not None
+            assert ver2 is not None
 
             # Create client and perform operations
             client = pydgraph.DgraphClient(client_stub)
@@ -153,9 +141,9 @@ class TestDgraphClientStubContextManager(helper.ClientIntegrationTestCase):
             txn = client.txn(read_only=True)
             query = "{ me(func: has(name)) { name } }"
             resp = txn.query(query)
-            self.assertIsNotNone(resp)
+            assert resp is not None
 
-    def test_context_manager_nested_with_client_operations(self):
+    def test_context_manager_nested_with_client_operations(self) -> None:
         """Test full workflow: stub context manager with client and transaction operations."""
         with pydgraph.DgraphClientStub(addr=self.TEST_SERVER_ADDR) as stub:
             client = pydgraph.DgraphClient(stub)
@@ -169,27 +157,25 @@ class TestDgraphClientStubContextManager(helper.ClientIntegrationTestCase):
             # Perform mutation and query
             with client.txn() as txn:
                 response = txn.mutate(set_obj={"test_name": "ContextManagerTest"})
-                self.assertEqual(1, len(response.uids))
-                uid = list(response.uids.values())[0]
+                assert len(response.uids) == 1
+                uid = next(iter(response.uids.values()))
 
             # Verify data was committed
-            query = """{{
+            query = f"""{{
                 me(func: uid("{uid}")) {{
                     test_name
                 }}
-            }}""".format(
-                uid=uid
-            )
+            }}"""
 
             with client.txn(read_only=True) as txn:
                 resp = txn.query(query)
                 import json
 
                 results = json.loads(resp.json).get("me")
-                self.assertEqual([{"test_name": "ContextManagerTest"}], results)
+                assert results == [{"test_name": "ContextManagerTest"}]
 
 
-def suite():
+def suite() -> unittest.TestSuite:
     """Returns a test suite object."""
     suite_obj = unittest.TestSuite()
     suite_obj.addTest(TestDgraphClientStub())
