@@ -3,9 +3,14 @@
 
 """Tests for transaction retry utilities."""
 
+from __future__ import annotations
+
 import asyncio
 import unittest
+from typing import Any
 from unittest.mock import MagicMock
+
+import pytest
 
 import pydgraph
 from pydgraph import errors
@@ -25,27 +30,27 @@ __maintainer__ = "Istari Digital, Inc. <dgraph-admin@istaridigital.com>"
 class TestCalculateDelay(unittest.TestCase):
     """Tests for delay calculation with exponential backoff."""
 
-    def test_initial_delay(self):
+    def test_initial_delay(self) -> None:
         """First attempt should use base delay."""
         delay = _calculate_delay(0, base_delay=0.1, max_delay=5.0, jitter=0.0)
-        self.assertEqual(delay, 0.1)
+        assert delay == 0.1
 
-    def test_exponential_backoff(self):
+    def test_exponential_backoff(self) -> None:
         """Delay should double with each attempt."""
         delay1 = _calculate_delay(0, base_delay=0.1, max_delay=5.0, jitter=0.0)
         delay2 = _calculate_delay(1, base_delay=0.1, max_delay=5.0, jitter=0.0)
         delay3 = _calculate_delay(2, base_delay=0.1, max_delay=5.0, jitter=0.0)
 
-        self.assertEqual(delay1, 0.1)
-        self.assertEqual(delay2, 0.2)
-        self.assertEqual(delay3, 0.4)
+        assert delay1 == 0.1
+        assert delay2 == 0.2
+        assert delay3 == 0.4
 
-    def test_max_delay_cap(self):
+    def test_max_delay_cap(self) -> None:
         """Delay should not exceed max_delay."""
         delay = _calculate_delay(100, base_delay=0.1, max_delay=5.0, jitter=0.0)
-        self.assertEqual(delay, 5.0)
+        assert delay == 5.0
 
-    def test_jitter_adds_randomness(self):
+    def test_jitter_adds_randomness(self) -> None:
         """Jitter should add some randomness to delay."""
         delays = set()
         for _ in range(10):
@@ -54,33 +59,33 @@ class TestCalculateDelay(unittest.TestCase):
 
         # With jitter, we should get different values
         # Base delay is 1.0, jitter 0.5 means delay is in range [1.0, 1.5]
-        self.assertTrue(all(1.0 <= d <= 1.5 for d in delays))
+        assert all(1.0 <= d <= 1.5 for d in delays)
 
 
 class TestIsRetriable(unittest.TestCase):
     """Tests for error classification."""
 
-    def test_aborted_error_is_retriable(self):
+    def test_aborted_error_is_retriable(self) -> None:
         """AbortedError should be retriable."""
         error = errors.AbortedError()
-        self.assertTrue(_is_retriable(error))
+        assert _is_retriable(error)
 
-    def test_retriable_error_is_retriable(self):
+    def test_retriable_error_is_retriable(self) -> None:
         """RetriableError should be retriable."""
         error = errors.RetriableError(Exception("test"))
-        self.assertTrue(_is_retriable(error))
+        assert _is_retriable(error)
 
-    def test_other_errors_not_retriable(self):
+    def test_other_errors_not_retriable(self) -> None:
         """Other errors should not be retriable."""
-        self.assertFalse(_is_retriable(Exception("test")))
-        self.assertFalse(_is_retriable(ValueError("test")))
-        self.assertFalse(_is_retriable(errors.TransactionError("test")))
+        assert not _is_retriable(Exception("test"))
+        assert not _is_retriable(ValueError("test"))
+        assert not _is_retriable(errors.TransactionError("test"))
 
 
 class TestRetryGenerator(unittest.TestCase):
     """Tests for sync retry generator."""
 
-    def test_success_on_first_attempt(self):
+    def test_success_on_first_attempt(self) -> None:
         """Should not retry if first attempt succeeds."""
         # Use very short delays to avoid actual sleeping
         attempts = 0
@@ -89,52 +94,52 @@ class TestRetryGenerator(unittest.TestCase):
                 attempts += 1
                 # Success - no exception
 
-        self.assertEqual(attempts, 1)
+        assert attempts == 1
 
-    def test_retry_on_aborted_error(self):
+    def test_retry_on_aborted_error(self) -> None:
         """Should retry on AbortedError."""
         attempts = 0
         for attempt in retry(max_retries=3, base_delay=0.001, jitter=0.0):
             with attempt:
                 attempts += 1
                 if attempts < 3:
-                    raise errors.AbortedError()
+                    raise errors.AbortedError
                 # Third attempt succeeds
 
-        self.assertEqual(attempts, 3)
+        assert attempts == 3
 
-    def test_raises_after_max_retries(self):
+    def test_raises_after_max_retries(self) -> None:
         """Should raise AbortedError after exhausting retries."""
         attempts = 0
 
-        with self.assertRaises(errors.AbortedError):
+        with pytest.raises(errors.AbortedError):
             for attempt in retry(max_retries=2, base_delay=0.001, jitter=0.0):
                 with attempt:
                     attempts += 1
-                    raise errors.AbortedError()
+                    raise errors.AbortedError
 
-        self.assertEqual(attempts, 3)  # Initial + 2 retries
+        assert attempts == 3  # Initial + 2 retries
 
-    def test_non_retriable_error_propagates(self):
+    def test_non_retriable_error_propagates(self) -> None:
         """Non-retriable errors should propagate immediately."""
         attempts = 0
 
-        with self.assertRaises(ValueError):
+        with pytest.raises(ValueError):
             for attempt in retry(max_retries=3, base_delay=0.001, jitter=0.0):
                 with attempt:
                     attempts += 1
                     raise ValueError("not retriable")
 
-        self.assertEqual(attempts, 1)
+        assert attempts == 1
 
 
 class TestRetryAsyncGenerator(unittest.TestCase):
     """Tests for async retry generator."""
 
-    def test_success_on_first_attempt(self):
+    def test_success_on_first_attempt(self) -> None:
         """Should not retry if first attempt succeeds."""
 
-        async def test():
+        async def test() -> int:
             attempts = 0
             async for attempt in retry_async(
                 max_retries=3, base_delay=0.001, jitter=0.0
@@ -144,13 +149,13 @@ class TestRetryAsyncGenerator(unittest.TestCase):
                     # Success
             return attempts
 
-        result = asyncio.get_event_loop().run_until_complete(test())
-        self.assertEqual(result, 1)
+        result = asyncio.run(test())
+        assert result == 1
 
-    def test_retry_on_aborted_error(self):
+    def test_retry_on_aborted_error(self) -> None:
         """Should retry on AbortedError."""
 
-        async def test():
+        async def test() -> int:
             attempts = 0
             async for attempt in retry_async(
                 max_retries=3, base_delay=0.001, jitter=0.0
@@ -158,183 +163,180 @@ class TestRetryAsyncGenerator(unittest.TestCase):
                 with attempt:
                     attempts += 1
                     if attempts < 3:
-                        raise errors.AbortedError()
+                        raise errors.AbortedError
             return attempts
 
-        result = asyncio.get_event_loop().run_until_complete(test())
-        self.assertEqual(result, 3)
+        result = asyncio.run(test())
+        assert result == 3
 
-    def test_raises_after_max_retries(self):
+    def test_raises_after_max_retries(self) -> None:
         """Should raise AbortedError after exhausting retries."""
 
-        async def test():
+        async def test() -> int:
             attempts = 0
             async for attempt in retry_async(
                 max_retries=2, base_delay=0.001, jitter=0.0
             ):
                 with attempt:
                     attempts += 1
-                    raise errors.AbortedError()
+                    raise errors.AbortedError
             return attempts
 
-        with self.assertRaises(errors.AbortedError):
-            asyncio.get_event_loop().run_until_complete(test())
+        with pytest.raises(errors.AbortedError):
+            asyncio.run(test())
 
 
 class TestWithRetryDecorator(unittest.TestCase):
     """Tests for sync retry decorator."""
 
-    def test_success_on_first_attempt(self):
+    def test_success_on_first_attempt(self) -> None:
         """Should not retry if function succeeds."""
 
         @with_retry(max_retries=3, base_delay=0.001, jitter=0.0)
-        def my_func():
+        def my_func() -> str:
             return "success"
 
         result = my_func()
-        self.assertEqual(result, "success")
+        assert result == "success"
 
-    def test_retry_on_aborted_error(self):
+    def test_retry_on_aborted_error(self) -> None:
         """Should retry on AbortedError."""
         attempts = {"count": 0}
 
         @with_retry(max_retries=3, base_delay=0.001, jitter=0.0)
-        def my_func():
+        def my_func() -> str:
             attempts["count"] += 1
             if attempts["count"] < 3:
-                raise errors.AbortedError()
+                raise errors.AbortedError
             return "success"
 
         result = my_func()
-        self.assertEqual(result, "success")
-        self.assertEqual(attempts["count"], 3)
+        assert result == "success"
+        assert attempts["count"] == 3
 
-    def test_raises_after_max_retries(self):
+    def test_raises_after_max_retries(self) -> None:
         """Should raise after exhausting retries."""
 
         @with_retry(max_retries=2, base_delay=0.001, jitter=0.0)
-        def my_func():
-            raise errors.AbortedError()
+        def my_func() -> None:
+            raise errors.AbortedError
 
-        with self.assertRaises(errors.AbortedError):
+        with pytest.raises(errors.AbortedError):
             my_func()
 
-    def test_preserves_function_metadata(self):
+    def test_preserves_function_metadata(self) -> None:
         """Should preserve function name and docstring."""
 
         @with_retry(base_delay=0.001, jitter=0.0)
-        def my_documented_func():
+        def my_documented_func() -> None:
             """This is a docstring."""
-            pass
 
-        self.assertEqual(my_documented_func.__name__, "my_documented_func")
-        self.assertEqual(my_documented_func.__doc__, "This is a docstring.")
+        assert my_documented_func.__name__ == "my_documented_func"
+        assert my_documented_func.__doc__ == "This is a docstring."
 
 
 class TestWithRetryAsyncDecorator(unittest.TestCase):
     """Tests for async retry decorator."""
 
-    def test_success_on_first_attempt(self):
+    def test_success_on_first_attempt(self) -> None:
         """Should not retry if function succeeds."""
 
         @with_retry_async(max_retries=3, base_delay=0.001, jitter=0.0)
-        async def my_func():
+        async def my_func() -> str:
             return "success"
 
-        result = asyncio.get_event_loop().run_until_complete(my_func())
-        self.assertEqual(result, "success")
+        result = asyncio.run(my_func())
+        assert result == "success"
 
-    def test_retry_on_aborted_error(self):
+    def test_retry_on_aborted_error(self) -> None:
         """Should retry on AbortedError."""
         attempts = {"count": 0}
 
         @with_retry_async(max_retries=3, base_delay=0.001, jitter=0.0)
-        async def my_func():
+        async def my_func() -> str:
             attempts["count"] += 1
             if attempts["count"] < 3:
-                raise errors.AbortedError()
+                raise errors.AbortedError
             return "success"
 
-        result = asyncio.get_event_loop().run_until_complete(my_func())
-        self.assertEqual(result, "success")
-        self.assertEqual(attempts["count"], 3)
+        result = asyncio.run(my_func())
+        assert result == "success"
+        assert attempts["count"] == 3
 
-    def test_raises_after_max_retries(self):
+    def test_raises_after_max_retries(self) -> None:
         """Should raise after exhausting retries."""
 
         @with_retry_async(max_retries=2, base_delay=0.001, jitter=0.0)
-        async def my_func():
-            raise errors.AbortedError()
+        async def my_func() -> None:
+            raise errors.AbortedError
 
-        with self.assertRaises(errors.AbortedError):
-            asyncio.get_event_loop().run_until_complete(my_func())
+        with pytest.raises(errors.AbortedError):
+            asyncio.run(my_func())
 
 
 class TestRetryImports(unittest.TestCase):
     """Tests for module-level imports."""
 
-    def test_retry_exported_from_pydgraph(self):
+    def test_retry_exported_from_pydgraph(self) -> None:
         """Retry utilities should be importable from pydgraph."""
-        self.assertTrue(hasattr(pydgraph, "retry"))
-        self.assertTrue(hasattr(pydgraph, "retry_async"))
-        self.assertTrue(hasattr(pydgraph, "with_retry"))
-        self.assertTrue(hasattr(pydgraph, "with_retry_async"))
-        self.assertTrue(hasattr(pydgraph, "run_transaction"))
-        self.assertTrue(hasattr(pydgraph, "run_transaction_async"))
+        assert hasattr(pydgraph, "retry")
+        assert hasattr(pydgraph, "retry_async")
+        assert hasattr(pydgraph, "with_retry")
+        assert hasattr(pydgraph, "with_retry_async")
+        assert hasattr(pydgraph, "run_transaction")
+        assert hasattr(pydgraph, "run_transaction_async")
 
 
 class TestRetryWithRetriableError(unittest.TestCase):
     """Tests for RetriableError handling."""
 
-    def test_retry_on_retriable_error(self):
+    def test_retry_on_retriable_error(self) -> None:
         """Should also retry on RetriableError."""
         attempts = {"count": 0}
 
         @with_retry(max_retries=3, base_delay=0.001, jitter=0.0)
-        def my_func():
+        def my_func() -> str:
             attempts["count"] += 1
             if attempts["count"] < 2:
                 raise errors.RetriableError(Exception("temporary"))
             return "success"
 
         result = my_func()
-        self.assertEqual(result, "success")
-        self.assertEqual(attempts["count"], 2)
+        assert result == "success"
+        assert attempts["count"] == 2
 
 
 class TestRetryExponentialBackoff(unittest.TestCase):
     """Tests to verify exponential backoff behavior with timing."""
 
-    def test_backoff_increases_delay(self):
+    def test_backoff_increases_delay(self) -> None:
         """Verify delays increase exponentially."""
         # Import the module directly using importlib (pydgraph.retry is shadowed by the function)
         import importlib
 
         retry_mod = importlib.import_module("pydgraph.retry")
 
-        sleep_durations = []
+        sleep_durations: list[float] = []
         original_sleep = retry_mod.time.sleep
 
-        def mock_sleep(duration):
+        def mock_sleep(duration: float) -> None:
             sleep_durations.append(duration)
             # Don't actually sleep
 
         # Temporarily replace sleep
-        retry_mod.time.sleep = mock_sleep
+        retry_mod.time.sleep = mock_sleep  # type: ignore[method-assign]
         try:
-            attempts = 0
-            with self.assertRaises(errors.AbortedError):
+            with pytest.raises(errors.AbortedError):
                 for attempt in retry(max_retries=3, base_delay=0.1, jitter=0.0):
                     with attempt:
-                        attempts += 1
-                        raise errors.AbortedError()
+                        raise errors.AbortedError
 
             # Should have 3 sleeps (between attempts 1-2, 2-3, 3-4)
-            self.assertEqual(len(sleep_durations), 3)
+            assert len(sleep_durations) == 3
             # Verify exponential backoff: 0.1, 0.2, 0.4
-            self.assertAlmostEqual(sleep_durations[0], 0.1, places=3)
-            self.assertAlmostEqual(sleep_durations[1], 0.2, places=3)
-            self.assertAlmostEqual(sleep_durations[2], 0.4, places=3)
+            assert sleep_durations[0] == pytest.approx(0.1, abs=0.001)
+            assert sleep_durations[1] == pytest.approx(0.2, abs=0.001)
+            assert sleep_durations[2] == pytest.approx(0.4, abs=0.001)
         finally:
             retry_mod.time.sleep = original_sleep
 
@@ -342,37 +344,37 @@ class TestRetryExponentialBackoff(unittest.TestCase):
 class TestParameterValidation(unittest.TestCase):
     """Tests for parameter validation."""
 
-    def test_negative_max_retries_raises(self):
+    def test_negative_max_retries_raises(self) -> None:
         """Negative max_retries should raise ValueError."""
         from pydgraph.retry import _validate_retry_params
 
-        with self.assertRaises(ValueError) as ctx:
+        with pytest.raises(ValueError) as ctx:
             _validate_retry_params(
                 max_retries=-1, base_delay=0.1, max_delay=5.0, jitter=0.1
             )
-        self.assertIn("max_retries", str(ctx.exception))
+        assert "max_retries" in str(ctx.value)
 
-    def test_negative_base_delay_raises(self):
+    def test_negative_base_delay_raises(self) -> None:
         """Negative base_delay should raise ValueError."""
         from pydgraph.retry import _validate_retry_params
 
-        with self.assertRaises(ValueError) as ctx:
+        with pytest.raises(ValueError) as ctx:
             _validate_retry_params(
                 max_retries=3, base_delay=-0.1, max_delay=5.0, jitter=0.1
             )
-        self.assertIn("base_delay", str(ctx.exception))
+        assert "base_delay" in str(ctx.value)
 
-    def test_invalid_jitter_raises(self):
+    def test_invalid_jitter_raises(self) -> None:
         """Jitter outside [0, 1] should raise ValueError."""
         from pydgraph.retry import _validate_retry_params
 
-        with self.assertRaises(ValueError) as ctx:
+        with pytest.raises(ValueError) as ctx:
             _validate_retry_params(
                 max_retries=3, base_delay=0.1, max_delay=5.0, jitter=1.5
             )
-        self.assertIn("jitter", str(ctx.exception))
+        assert "jitter" in str(ctx.value)
 
-    def test_valid_params_pass(self):
+    def test_valid_params_pass(self) -> None:
         """Valid parameters should not raise."""
         from pydgraph.retry import _validate_retry_params
 
@@ -386,7 +388,7 @@ class TestParameterValidation(unittest.TestCase):
 class TestRunTransaction(unittest.TestCase):
     """Tests for run_transaction helper."""
 
-    def test_success_on_first_attempt(self):
+    def test_success_on_first_attempt(self) -> None:
         """Should succeed if operation doesn't raise."""
         from pydgraph.retry import run_transaction
 
@@ -394,16 +396,16 @@ class TestRunTransaction(unittest.TestCase):
         mock_txn = MagicMock()
         mock_client.txn.return_value = mock_txn
 
-        def operation(txn):
+        def operation(txn: Any) -> str:
             return "success"
 
         result = run_transaction(
             mock_client, operation, max_retries=3, base_delay=0.001
         )
-        self.assertEqual(result, "success")
+        assert result == "success"
         mock_txn.discard.assert_called_once()
 
-    def test_retry_on_aborted_error(self):
+    def test_retry_on_aborted_error(self) -> None:
         """Should retry on AbortedError."""
         from pydgraph.retry import run_transaction
 
@@ -412,19 +414,19 @@ class TestRunTransaction(unittest.TestCase):
         mock_client.txn.return_value = mock_txn
         attempts = {"count": 0}
 
-        def operation(txn):
+        def operation(txn: Any) -> str:
             attempts["count"] += 1
             if attempts["count"] < 3:
-                raise errors.AbortedError()
+                raise errors.AbortedError
             return "success"
 
         result = run_transaction(
             mock_client, operation, max_retries=5, base_delay=0.001, jitter=0.0
         )
-        self.assertEqual(result, "success")
-        self.assertEqual(attempts["count"], 3)
+        assert result == "success"
+        assert attempts["count"] == 3
 
-    def test_raises_after_max_retries(self):
+    def test_raises_after_max_retries(self) -> None:
         """Should raise after exhausting retries."""
         from pydgraph.retry import run_transaction
 
@@ -432,16 +434,16 @@ class TestRunTransaction(unittest.TestCase):
         mock_txn = MagicMock()
         mock_client.txn.return_value = mock_txn
 
-        def operation(txn):
-            raise errors.AbortedError()
+        def operation(txn: Any) -> None:
+            raise errors.AbortedError
 
-        with self.assertRaises(errors.AbortedError):
+        with pytest.raises(errors.AbortedError):
             run_transaction(
                 mock_client, operation, max_retries=2, base_delay=0.001, jitter=0.0
             )
 
 
-def suite():
+def suite() -> unittest.TestSuite:
     """Returns a test suite object."""
     suite_obj = unittest.TestSuite()
     suite_obj.addTest(unittest.makeSuite(TestCalculateDelay))
