@@ -1,5 +1,7 @@
-# SPDX-FileCopyrightText: © 2017-2025 Istari Digital, Inc.
+# SPDX-FileCopyrightText: © 2017-2026 Istari Digital, Inc.
 # SPDX-License-Identifier: Apache-2.0
+
+from __future__ import annotations
 
 __author__ = "Garvit Pahal"
 __maintainer__ = "Istari Digital, Inc. <dgraph-admin@istaridigital.com>"
@@ -9,172 +11,166 @@ import logging
 import time
 import unittest
 
+import pytest
+
 import pydgraph
 from tests import helper
 
 
 class TestTxn(helper.ClientIntegrationTestCase):
-    def setUp(self):
-        super(TestTxn, self).setUp()
+    def setUp(self) -> None:
+        super().setUp()
 
         helper.drop_all(self.client)
         helper.set_schema(self.client, "name: string @index(fulltext) @upsert .")
 
-    def test_query_after_commit(self):
+    def test_query_after_commit(self) -> None:
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
 
         txn.commit()
 
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
 
-        with self.assertRaises(Exception):  # noqa: B017
+        with pytest.raises(
+            Exception, match="Transaction has already been committed or discarded"
+        ):
             txn.query(query)
 
-    def test_mutate_after_commit(self):
+    def test_mutate_after_commit(self) -> None:
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
         txn.commit()
 
-        with self.assertRaises(Exception):  # noqa: B017
+        with pytest.raises(
+            Exception, match="Transaction has already been committed or discarded"
+        ):
             txn.mutate(set_obj={"name": "Manish2"})
 
-    def test_commit_now(self):
+    def test_commit_now(self) -> None:
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"}, commit_now=True)
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
-        for _, uid in response.uids.items():
+        assert len(response.uids) == 1, "Nothing was assigned"
+        for uid in response.uids.values():
             uid = uid
 
-        self.assertRaises(Exception, txn.commit)  # noqa: B017
+        self.assertRaises(Exception, txn.commit)
 
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
         resp = self.client.txn(read_only=True).query(query)
-        self.assertEqual([{"name": "Manish"}], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == [{"name": "Manish"}]
 
-    def test_discard(self):
+    def test_discard(self) -> None:
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
         txn.commit()
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
 
         txn2 = self.client.txn()
         _ = txn2.mutate(set_obj={"uid": uid, "name": "Manish2"})
 
         txn.discard()
-        self.assertRaises(Exception, txn.commit)  # noqa: B017
+        self.assertRaises(Exception, txn.commit)
 
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
         resp = self.client.txn(read_only=True).query(query)
-        self.assertEqual([{"name": "Manish"}], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == [{"name": "Manish"}]
 
-    def test_mutate_error(self):
+    def test_mutate_error(self) -> None:
         txn = self.client.txn()
-        with self.assertRaises(Exception):  # noqa: B017
+        with pytest.raises(Exception):
             # Following N-Quad is invalid
             _ = txn.mutate(set_nquads="_:node <name> Manish")
 
-        self.assertRaises(Exception, txn.commit)  # noqa: B017
+        self.assertRaises(Exception, txn.commit)
 
-    def test_read_at_start_ts(self):
+    def test_read_at_start_ts(self) -> None:
         """Tests read after write when readTs == startTs"""
 
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
 
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
         resp = txn.query(query)
-        self.assertEqual([{"name": "Manish"}], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == [{"name": "Manish"}]
 
-    def test_read_before_start_ts(self):
+    def test_read_before_start_ts(self) -> None:
         """Tests read before write when readTs < startTs"""
 
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
 
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
 
         resp = self.client.txn(read_only=True).query(query)
-        self.assertEqual([], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == []
 
-    def test_read_after_start_ts(self):
+    def test_read_after_start_ts(self) -> None:
         """Tests read after committing a write when readTs > startTs"""
 
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
         txn.commit()
 
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
 
         resp = self.client.txn(read_only=True).query(query)
-        self.assertEqual([{"name": "Manish"}], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == [{"name": "Manish"}]
 
-    def test_read_before_and_after_start_ts(self):
+    def test_read_before_and_after_start_ts(self) -> None:
         """Test read before and after committing a transaction when
         readTs1 < startTs and readTs2 > startTs"""
 
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
         txn.commit()
 
@@ -184,58 +180,54 @@ class TestTxn(helper.ClientIntegrationTestCase):
         txn3 = self.client.txn()
         _ = txn3.mutate(set_obj={"uid": uid, "name": "Manish2"})
 
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
 
         # object is unchanged since txn3 is uncommitted
         resp2 = txn2.query(query)
-        self.assertEqual([{"name": "Manish"}], json.loads(resp2.json).get("me"))
+        assert json.loads(resp2.json).get("me") == [{"name": "Manish"}]
 
         # once txn3 is committed, other txns observe the update
         txn3.commit()
 
         resp4 = self.client.txn(read_only=True).query(query)
-        self.assertEqual([{"name": "Manish2"}], json.loads(resp4.json).get("me"))
+        assert json.loads(resp4.json).get("me") == [{"name": "Manish2"}]
 
-    def test_read_from_new_client(self):
+    def test_read_from_new_client(self) -> None:
         """Tests committed reads from a new client with startTs == 0."""
 
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
         txn.commit()
 
         client2 = helper.create_client(self.TEST_SERVER_ADDR)
         client2.login("groot", "password")
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
 
         resp2 = client2.txn(read_only=True).query(query)
-        self.assertEqual([{"name": "Manish"}], json.loads(resp2.json).get("me"))
-        self.assertTrue(resp2.txn.start_ts > 0)
+        assert json.loads(resp2.json).get("me") == [{"name": "Manish"}]
+        assert resp2.txn.start_ts > 0
 
         txn2 = client2.txn()
         response = txn2.mutate(set_obj={"uid": uid, "name": "Manish2"})
-        self.assertTrue(response.txn.start_ts > 0)
+        assert response.txn.start_ts > 0
         txn2.commit()
 
         resp = self.client.txn(read_only=True).query(query)
-        self.assertEqual([{"name": "Manish2"}], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == [{"name": "Manish2"}]
 
-    def test_read_only_txn(self):
+    def test_read_only_txn(self) -> None:
         """Tests read-only transactions. Read-only transactions should
         not advance the start ts nor should allow mutations or commits.
 
@@ -259,10 +251,8 @@ class TestTxn(helper.ClientIntegrationTestCase):
         # Allow small timestamp differences due to v23+ rollup behavior
         # In most cases timestamps should be equal, but rollups may cause
         # small increments in CI environments
-        self.assertLessEqual(
-            abs(start_ts1 - start_ts2),
-            5,
-            "Timestamps should be equal or differ by at most 5 due to rollups",
+        assert abs(start_ts1 - start_ts2) <= 5, (
+            "Timestamps should be equal or differ by at most 5 due to rollups"
         )
 
         txn = self.client.txn(read_only=True)
@@ -272,26 +262,32 @@ class TestTxn(helper.ClientIntegrationTestCase):
         start_ts2 = resp2.txn.start_ts
 
         # Within the same transaction, timestamps should always be equal
-        self.assertEqual(start_ts1, start_ts2)
+        assert start_ts1 == start_ts2
 
-        with self.assertRaises(Exception):  # noqa: B017
+        with pytest.raises(
+            Exception, match="Readonly transaction cannot run mutations"
+        ):
             txn.mutate(set_obj={"name": "Manish"})
-        with self.assertRaises(Exception):  # noqa: B017
+        with pytest.raises(
+            Exception, match="Readonly transaction cannot run mutations"
+        ):
             txn.commit()
 
-    def test_best_effort_txn(self):
+    def test_best_effort_txn(self) -> None:
         """Tests best-effort transactions."""
 
         helper.drop_all(self.client)
         helper.set_schema(self.client, "name: string @index(exact) .")
 
-        with self.assertRaises(Exception):  # noqa: B017
+        with pytest.raises(
+            Exception, match="Best effort transactions are only compatible"
+        ):
             self.client.txn(read_only=False, best_effort=True)
 
         query = "{ me(func: has(name)) {name} }"
         rtxn = self.client.txn(read_only=True, best_effort=True)
         resp = rtxn.query(query)
-        self.assertEqual([], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == []
 
         txn = self.client.txn()
         resp = txn.mutate(set_obj={"name": "Manish"})
@@ -299,26 +295,26 @@ class TestTxn(helper.ClientIntegrationTestCase):
         mu_ts = resp.txn.commit_ts
 
         resp = rtxn.query(query)
-        self.assertEqual([], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == []
 
         while True:
             txn = self.client.txn(read_only=True)
             resp = txn.query(query)
             if resp.txn.start_ts < mu_ts:
                 continue
-            self.assertEqual([{"name": "Manish"}], json.loads(resp.json).get("me"))
+            assert json.loads(resp.json).get("me") == [{"name": "Manish"}]
             break
 
-    def test_conflict(self):
+    def test_conflict(self) -> None:
         """Tests committing two transactions which conflict."""
 
         helper.drop_all(self.client)
 
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
 
         txn2 = self.client.txn()
@@ -328,39 +324,35 @@ class TestTxn(helper.ClientIntegrationTestCase):
         self.assertRaises(pydgraph.AbortedError, txn2.commit)
 
         txn3 = self.client.txn()
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
 
         resp3 = txn3.query(query)
-        self.assertEqual([{"name": "Manish"}], json.loads(resp3.json).get("me"))
+        assert json.loads(resp3.json).get("me") == [{"name": "Manish"}]
 
-    def test_conflict_reverse_order(self):
+    def test_conflict_reverse_order(self) -> None:
         """Tests committing a transaction after a newer transaction has been
         committed."""
 
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
 
         txn2 = self.client.txn()
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
 
         resp = txn2.query(query)
-        self.assertEqual([], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == []
 
         _ = txn2.mutate(set_obj={"uid": uid, "name": "Jan the man"})
         txn2.commit()
@@ -369,14 +361,14 @@ class TestTxn(helper.ClientIntegrationTestCase):
 
         txn3 = self.client.txn()
         resp = txn3.query(query)
-        self.assertEqual([{"name": "Jan the man"}], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == [{"name": "Jan the man"}]
 
-    def test_commit_conflict(self):
+    def test_commit_conflict(self) -> None:
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
 
         txn2 = self.client.txn()
@@ -389,23 +381,21 @@ class TestTxn(helper.ClientIntegrationTestCase):
         _ = txn3.mutate(set_obj={"uid": uid, "name": "Jan the man"})
         txn3.commit()
 
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid:s}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
 
         resp4 = self.client.txn(read_only=True).query(query)
-        self.assertEqual([{"name": "Jan the man"}], json.loads(resp4.json).get("me"))
+        assert json.loads(resp4.json).get("me") == [{"name": "Jan the man"}]
 
-    def test_mutate_conflict(self):
+    def test_mutate_conflict(self) -> None:
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
 
         txn2 = self.client.txn()
@@ -416,7 +406,7 @@ class TestTxn(helper.ClientIntegrationTestCase):
         _ = txn.mutate(set_obj={"uid": uid, "name": "Manish2"})
         self.assertRaises(pydgraph.AbortedError, txn.commit)
 
-    def test_read_index_key_same_txn(self):
+    def test_read_index_key_same_txn(self) -> None:
         """Tests reading an indexed field within a transaction. The read
         should return the results from before any writes of the same
         txn."""
@@ -426,9 +416,9 @@ class TestTxn(helper.ClientIntegrationTestCase):
 
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Manish"})
-        self.assertEqual(1, len(response.uids), "Nothing was assigned")
+        assert len(response.uids) == 1, "Nothing was assigned"
 
-        for _, uid in response.uids.items():
+        for uid in response.uids.values():
             uid = uid
 
         query = """{
@@ -438,11 +428,9 @@ class TestTxn(helper.ClientIntegrationTestCase):
         }"""
 
         resp = txn.query(query)
-        self.assertEqual(
-            [], json.loads(resp.json).get("me"), "Expected 0 nodes read from index"
-        )
+        assert json.loads(resp.json).get("me") == [], "Expected 0 nodes read from index"
 
-    def test_non_string_variable(self):
+    def test_non_string_variable(self) -> None:
         """Tests sending a variable map with non-string values or keys results
         in an Exception."""
         helper.drop_all(self.client)
@@ -458,17 +446,19 @@ class TestTxn(helper.ClientIntegrationTestCase):
             }
         """
         variables = {"$a": 1234}
-        with self.assertRaises(Exception):  # noqa: B017
-            _ = txn.query(query, variables=variables)
+        with pytest.raises(Exception):
+            _ = txn.query(query, variables=variables)  # type: ignore[arg-type]
 
-    def test_finished(self):
+    def test_finished(self) -> None:
         txn = self.client.txn()
         txn.mutate(set_nquads='_:animesh <name> "Animesh" .', commit_now=True)
 
-        with self.assertRaises(Exception):  # noqa: B017
+        with pytest.raises(
+            Exception, match="Transaction has already been committed or discarded"
+        ):
             txn.mutate(set_nquads='_:aman <name> "Aman" .', commit_now=True)
 
-    def test_mutate_facet(self):
+    def test_mutate_facet(self) -> None:
         """Tests mutations that include facets work as expected."""
         helper.drop_all(self.client)
         helper.set_schema(
@@ -503,24 +493,23 @@ _:a <friend> _:b (close_friend=true).
 """
         txn = self.client.txn()
         resp = txn.query(query)
-        self.assertEqual(
-            [{"name": "aaa", "name|close_friend": True}, {"name": "bbb"}],
-            json.loads(resp.json).get("q1"),
-        )
-        self.assertEqual(
-            [{"friend": {"name": "bbb", "friend|close_friend": True}}],
-            json.loads(resp.json).get("q2"),
-        )
+        assert json.loads(resp.json).get("q1") == [
+            {"name": "aaa", "name|close_friend": True},
+            {"name": "bbb"},
+        ]
+        assert json.loads(resp.json).get("q2") == [
+            {"friend": {"name": "bbb", "friend|close_friend": True}}
+        ]
 
 
 class TestSPStar(helper.ClientIntegrationTestCase):
-    def setUp(self):
-        super(TestSPStar, self).setUp()
+    def setUp(self) -> None:
+        super().setUp()
 
         helper.drop_all(self.client)
         helper.set_schema(self.client, "friend: [uid] .")
 
-    def test_sp_star(self):
+    def test_sp_star(self) -> None:
         """Tests a Subject Predicate Star query."""
 
         txn = self.client.txn()
@@ -528,13 +517,13 @@ class TestSPStar(helper.ClientIntegrationTestCase):
             set_obj={"uid": "_:manish", "name": "Manish", "friend": [{"name": "Jan"}]}
         )
         uid1 = response.uids["manish"]
-        self.assertEqual(2, len(response.uids), "Expected 2 nodes to be created")
+        assert len(response.uids) == 2, "Expected 2 nodes to be created"
 
         txn.commit()
 
         txn2 = self.client.txn()
         response2 = txn2.mutate(del_obj={"uid": uid1, "friend": None})
-        self.assertEqual(0, len(response2.uids))
+        assert len(response2.uids) == 0
 
         response3 = txn2.mutate(
             set_obj={
@@ -543,28 +532,25 @@ class TestSPStar(helper.ClientIntegrationTestCase):
                 "friend": [{"uid": "_:jan2", "name": "Jan2"}],
             }
         )
-        self.assertEqual(1, len(response3.uids))
+        assert len(response3.uids) == 1
         uid2 = response3.uids["jan2"]
 
-        query = """{{
-            me(func: uid("{uid:s}")) {{
+        query = f"""{{
+            me(func: uid("{uid1:s}")) {{
                 uid
                 friend {{
                     uid
                     name
                 }}
             }}
-        }}""".format(
-            uid=uid1
-        )
+        }}"""
 
         resp = txn2.query(query)
-        self.assertEqual(
-            [{"uid": uid1, "friend": [{"name": "Jan2", "uid": uid2}]}],
-            json.loads(resp.json).get("me"),
-        )
+        assert [{"uid": uid1, "friend": [{"name": "Jan2", "uid": uid2}]}] == json.loads(
+            resp.json
+        ).get("me")
 
-    def test_sp_star2(self):
+    def test_sp_star2(self) -> None:
         """Second test of Subject Predicate Star"""
 
         txn = self.client.txn()
@@ -575,32 +561,29 @@ class TestSPStar(helper.ClientIntegrationTestCase):
                 "friend": [{"uid": "_:jan", "name": "Jan"}],
             }
         )
-        self.assertEqual(2, len(response.uids))
+        assert len(response.uids) == 2
         uid1, uid2 = response.uids["manish"], response.uids["jan"]
 
-        query = """{{
-            me(func: uid("{uid:s}")) {{
+        query = f"""{{
+            me(func: uid("{uid1:s}")) {{
                 uid
                 friend {{
                     uid
                     name
                 }}
             }}
-        }}""".format(
-            uid=uid1
-        )
+        }}"""
 
         resp = txn.query(query)
-        self.assertEqual(
-            [{"uid": uid1, "friend": [{"name": "Jan", "uid": uid2}]}],
-            json.loads(resp.json).get("me"),
-        )
+        assert [{"uid": uid1, "friend": [{"name": "Jan", "uid": uid2}]}] == json.loads(
+            resp.json
+        ).get("me")
 
         deleted = txn.mutate(del_obj={"uid": uid1, "friend": None})
-        self.assertEqual(0, len(deleted.uids))
+        assert len(deleted.uids) == 0
 
         resp = txn.query(query)
-        self.assertEqual([{"uid": uid1}], json.loads(resp.json).get("me"))
+        assert [{"uid": uid1}] == json.loads(resp.json).get("me")
 
         # add an edge to Jan2
         response2 = txn.mutate(
@@ -610,28 +593,27 @@ class TestSPStar(helper.ClientIntegrationTestCase):
                 "friend": [{"uid": "_:jan2", "name": "Jan2"}],
             }
         )
-        self.assertEqual(1, len(response2.uids))
+        assert len(response2.uids) == 1
         uid2 = response2.uids["jan2"]
 
         resp = txn.query(query)
-        self.assertEqual(
-            [{"uid": uid1, "friend": [{"name": "Jan2", "uid": uid2}]}],
-            json.loads(resp.json).get("me"),
-        )
+        assert [{"uid": uid1, "friend": [{"name": "Jan2", "uid": uid2}]}] == json.loads(
+            resp.json
+        ).get("me")
 
         deleted2 = txn.mutate(del_obj={"uid": uid1, "friend": None})
-        self.assertEqual(0, len(deleted2.uids))
+        assert len(deleted2.uids) == 0
         resp = txn.query(query)
-        self.assertEqual([{"uid": uid1}], json.loads(resp.json).get("me"))
+        assert [{"uid": uid1}] == json.loads(resp.json).get("me")
 
 
 class TestContextManager(helper.ClientIntegrationTestCase):
-    def setUp(self):
-        super(TestContextManager, self).setUp()
+    def setUp(self) -> None:
+        super().setUp()
         helper.drop_all(self.client)
         helper.set_schema(self.client, "name: string @index(fulltext) .")
 
-    def test_context_manager_by_contextlib(self):
+    def test_context_manager_by_contextlib(self) -> None:
         """Test context manager via client.begin() for read-only transactions."""
         q = """
         {
@@ -642,10 +624,10 @@ class TestContextManager(helper.ClientIntegrationTestCase):
         """
         with self.client.begin(read_only=True, best_effort=True) as tx:
             response = tx.query(q)
-        self.assertIsNotNone(response)
+        assert response is not None
         _data = json.loads(response.json)
 
-    def test_context_manager_by_class(self):
+    def test_context_manager_by_class(self) -> None:
         """Test context manager using Txn class directly for read-only transactions."""
         q = """
         {
@@ -656,59 +638,54 @@ class TestContextManager(helper.ClientIntegrationTestCase):
         """
         with pydgraph.Txn(self.client, read_only=True, best_effort=True) as tx:
             response = tx.query(q)
-        self.assertIsNotNone(response)
+        assert response is not None
         _data = json.loads(response.json)
 
-    def test_context_manager_auto_commit(self):
+    def test_context_manager_auto_commit(self) -> None:
         """Test that write transactions automatically commit on successful completion."""
         with self.client.txn() as txn:
             response = txn.mutate(set_obj={"name": "Alice"})
-            self.assertEqual(1, len(response.uids), "Nothing was assigned")
-            uid = list(response.uids.values())[0]
+            assert len(response.uids) == 1, "Nothing was assigned"
+            uid = next(iter(response.uids.values()))
 
         # Verify the data was committed by querying in a new transaction
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
 
         resp = self.client.txn(read_only=True).query(query)
-        self.assertEqual([{"name": "Alice"}], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == [{"name": "Alice"}]
 
-    def test_context_manager_read_only_auto_discard(self):
+    def test_context_manager_read_only_auto_discard(self) -> None:
         """Test that read-only transactions automatically discard."""
         # Create some data first
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Bob"})
-        uid = list(response.uids.values())[0]
+        uid = next(iter(response.uids.values()))
         txn.commit()
 
         # Read-only transaction should auto-discard (not commit)
-        query = """{{
+        query = f"""{{
             me(func: uid("{uid}")) {{
                 name
             }}
-        }}""".format(
-            uid=uid
-        )
+        }}"""
 
         with self.client.txn(read_only=True) as txn:
             resp = txn.query(query)
-            self.assertEqual([{"name": "Bob"}], json.loads(resp.json).get("me"))
+            assert json.loads(resp.json).get("me") == [{"name": "Bob"}]
 
         # Transaction should be finished after context manager exits
-        self.assertTrue(txn._finished)
+        assert txn._finished
 
-    def test_context_manager_exception_handling(self):
+    def test_context_manager_exception_handling(self) -> None:
         """Test that exceptions cause automatic discard and are re-raised."""
-        with self.assertRaises(ValueError):
-            with self.client.txn() as txn:
-                response = txn.mutate(set_obj={"name": "Charlie"})
-                _uid = list(response.uids.values())[0]
-                raise ValueError("Test exception")
+        with self.assertRaises(ValueError), self.client.txn() as txn:
+            response = txn.mutate(set_obj={"name": "Charlie"})
+            _uid = next(iter(response.uids.values()))
+            raise ValueError("Test exception")
 
         # Verify transaction was discarded - data should not exist
         query = """{
@@ -722,29 +699,29 @@ class TestContextManager(helper.ClientIntegrationTestCase):
         # Should be empty or not contain Charlie
         if results:
             names = [r.get("name") for r in results]
-            self.assertNotIn("Charlie", names)
+            assert "Charlie" not in names
 
-    def test_context_manager_transaction_finished_after_exit(self):
+    def test_context_manager_transaction_finished_after_exit(self) -> None:
         """Test that transaction is marked as finished after exiting context manager."""
         with self.client.txn() as txn:
             txn.mutate(set_obj={"name": "David"})
-            self.assertFalse(txn._finished)
+            assert not txn._finished
 
         # Should be finished after exit
-        self.assertTrue(txn._finished)
+        assert txn._finished
 
         # Should not be able to use transaction after context manager
-        with self.assertRaises(Exception):  # noqa: B017
+        with self.assertRaises(Exception):
             txn.query("{ me() {} }")
 
-    def test_context_manager_multiple_mutations(self):
+    def test_context_manager_multiple_mutations(self) -> None:
         """Test multiple mutations within a single context manager."""
         with self.client.txn() as txn:
             response1 = txn.mutate(set_obj={"name": "Eve"})
-            _uid1 = list(response1.uids.values())[0]
+            _uid1 = next(iter(response1.uids.values()))
 
             response2 = txn.mutate(set_obj={"name": "Frank"})
-            _uid2 = list(response2.uids.values())[0]
+            _uid2 = next(iter(response2.uids.values()))
 
         # Verify both mutations were committed
         query = """{
@@ -756,64 +733,63 @@ class TestContextManager(helper.ClientIntegrationTestCase):
         resp = self.client.txn(read_only=True).query(query)
         results = json.loads(resp.json).get("me")
         names = [r.get("name") for r in results]
-        self.assertIn("Eve", names)
-        self.assertIn("Frank", names)
+        assert "Eve" in names
+        assert "Frank" in names
 
-    def test_context_manager_query_and_mutate(self):
+    def test_context_manager_query_and_mutate(self) -> None:
         """Test both query and mutate operations within a context manager."""
         # Create initial data
         txn = self.client.txn()
         response = txn.mutate(set_obj={"name": "Grace"})
-        uid = list(response.uids.values())[0]
+        uid = next(iter(response.uids.values()))
         txn.commit()
 
         # Query and update in context manager
         with self.client.txn() as txn:
-            query = """{{
+            query = f"""{{
                 me(func: uid("{uid}")) {{
                     name
                 }}
-            }}""".format(
-                uid=uid
-            )
+            }}"""
 
             resp = txn.query(query)
-            self.assertEqual([{"name": "Grace"}], json.loads(resp.json).get("me"))
+            assert json.loads(resp.json).get("me") == [{"name": "Grace"}]
 
             # Update the name
             txn.mutate(set_obj={"uid": uid, "name": "Grace Updated"})
 
         # Verify the update was committed
         resp = self.client.txn(read_only=True).query(query)
-        self.assertEqual([{"name": "Grace Updated"}], json.loads(resp.json).get("me"))
+        assert json.loads(resp.json).get("me") == [{"name": "Grace Updated"}]
 
-    def test_context_manager_invalid_nquad_exception(self):
+    def test_context_manager_invalid_nquad_exception(self) -> None:
         """Test that invalid operations cause proper exception handling and discard."""
-        with self.assertRaises(Exception):  # noqa: B017
-            with self.client.txn() as txn:
-                # This should fail with invalid N-Quad syntax
-                txn.mutate(set_nquads="_:node <name> InvalidWithoutQuotes")
+        with self.assertRaises(Exception), self.client.txn() as txn:
+            # This should fail with invalid N-Quad syntax
+            txn.mutate(set_nquads="_:node <name> InvalidWithoutQuotes")
 
         # Transaction should be finished
-        self.assertTrue(txn._finished)
+        assert txn._finished
 
-    def test_context_manager_read_only_cannot_mutate(self):
+    def test_context_manager_read_only_cannot_mutate(self) -> None:
         """Test that read-only transactions cannot mutate within context manager."""
-        with self.assertRaises(Exception):  # noqa: B017
-            with self.client.txn(read_only=True) as txn:
-                txn.mutate(set_obj={"name": "Should Fail"})
+        with (
+            pytest.raises(Exception),
+            self.client.txn(read_only=True) as txn,
+        ):
+            txn.mutate(set_obj={"name": "Should Fail"})
 
-    def test_context_manager_no_mutations_auto_commit(self):
+    def test_context_manager_no_mutations_auto_commit(self) -> None:
         """Test that transactions with no mutations don't error on auto-commit."""
         with self.client.txn() as txn:
             query = "{ me(func: has(name)) { name } }"
             _resp = txn.query(query)
 
         # Should complete without errors even though no mutations were made
-        self.assertTrue(txn._finished)
+        assert txn._finished
 
 
-def suite():
+def suite() -> unittest.TestSuite:
     s = unittest.TestSuite()
     s.addTest(TestTxn())
     s.addTest(TestSPStar())
