@@ -1,7 +1,5 @@
-# SPDX-FileCopyrightText: © 2017-2025 Istari Digital, Inc.
+# SPDX-FileCopyrightText: © 2017-2026 Istari Digital, Inc.
 # SPDX-License-Identifier: Apache-2.0
-
-# ruff: noqa: F401, F403
 
 # Async client
 from pydgraph.async_client import AsyncDgraphClient, async_open
@@ -9,9 +7,14 @@ from pydgraph.async_client_stub import AsyncDgraphClientStub
 from pydgraph.async_txn import AsyncTxn
 
 # Sync client
-from pydgraph.client import *
-from pydgraph.client_stub import *
-from pydgraph.errors import *
+from pydgraph.client import DgraphClient, open  # noqa: A004
+from pydgraph.client_stub import DgraphClientStub, client_stub
+from pydgraph.errors import (
+    AbortedError,
+    ConnectionError,  # noqa: A004
+    RetriableError,
+    TransactionError,
+)
 from pydgraph.proto.api_pb2 import (
     Check,
     Facet,
@@ -36,4 +39,38 @@ from pydgraph.retry import (
     with_retry,
     with_retry_async,
 )
-from pydgraph.txn import *
+from pydgraph.txn import Txn
+
+__all__ = [
+    "AbortedError",
+    "AsyncDgraphClient",
+    "AsyncDgraphClientStub",
+    "AsyncTxn",
+    "Check",
+    "ConnectionError",
+    "DgraphClient",
+    "DgraphClientStub",
+    "Facet",
+    "Latency",
+    "Mutation",
+    "NQuad",
+    "Operation",
+    "Payload",
+    "Request",
+    "Response",
+    "RetriableError",
+    "TransactionError",
+    "Txn",
+    "TxnContext",
+    "Value",
+    "Version",
+    "async_open",
+    "client_stub",
+    "open",
+    "retry",
+    "retry_async",
+    "run_transaction",
+    "run_transaction_async",
+    "with_retry",
+    "with_retry_async",
+]
