@@ -1,7 +1,11 @@
-# SPDX-FileCopyrightText: © 2017-2025 Istari Digital, Inc.
+# SPDX-FileCopyrightText: © 2017-2026 Istari Digital, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
 """Various utility functions."""
+
+from __future__ import annotations
+
+from typing import Any
 
 import grpc
 import grpc.aio
@@ -14,16 +18,17 @@ __version__ = VERSION
 __status__ = "development"
 
 
-def is_string(string):
+def is_string(string: Any) -> bool:
     """Checks if argument is a string."""
     return isinstance(string, str)
 
 
-def is_jwt_expired(exception):
+def is_jwt_expired(exception: Exception) -> bool:
+    """Returns true if the JWT token in the exception is expired."""
     return "Token is expired" in str(exception)
 
 
-def is_aborted_error(error):
+def is_aborted_error(error: Any) -> bool:
     """Returns true if the error is due to an aborted transaction.
 
     Uses duck typing to check for gRPC errors, which works with both
@@ -41,13 +46,13 @@ def is_aborted_error(error):
     return False
 
 
-def is_retriable_error(error):
+def is_retriable_error(error: Exception) -> bool:
     """Returns true if the error is retriable (e.g server is not ready yet)."""
     msg = str(error)
     return "Please retry" in msg or "opIndexing is already running" in msg
 
 
-def is_connection_error(error):
+def is_connection_error(error: Exception) -> bool:
     """Returns true if the error is caused connection issues."""
     msg = str(error)
     return "Unhealthy connection" in msg or "No connection exists" in msg
